pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "choreopy"
version = "0.1.0"
description = "Utility package for handling CRDs and resources in Python projects."
authors = [{ name = "Wim Henderickx", email = "wim.henderickx@gmail.com" }]
dependencies = [
    "annotated-types==0.7.0",
    "inflect==7.4.0",
    "more-itertools==10.5.0",
    "pydantic==2.9.2",
    "pydantic_core==2.23.4",
    "PyYAML==6.0.2",
    "typeguard==4.4.1",
    "typing_extensions==4.12.2"
]

#[tool.setuptools]
#packages = ["src"]
#py-modules = ["__init__"] # dash, not underscore
